allow consumers to signalise finished consumption
this allows consumers to signalise that the consumption is finished and offsets are marked, and thus the current rebalance can proceed without waiting for the DwellTime timeout.

this is useful in case the DwellTime is large enough to allow long synchronisation to external storage, but at the same time allow fast rebalances.